bsp: u-boot-fio: rpi: drop of_board
Drop OF_BOARD and move back to the default OF_EMBED as u-boot might not
necessarily be able to understand and parse the device tree produced by
the boot firmware (latest firmware + kernel 5.4 dtb breaks u-boot when
using OF_BOARD).

It is fine to use OF_EMBED as we now have multiple builds and machines
for the respective rpi variants.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>